Move room deletion event to room layout
